Game's state and state manager. Draw something on screen for play state.
This in turns
- added GameState and GameStateManager to control each state of the game
- Play state renders a testing text on screen
- SpriteBatch, and OrthographicCamera variables are cached in GameState
  class for easy usage in its code, and reduce referring performance hit
  thus no need to go through Game class to get them every single time
  (this is the reason why he did it in 1st tutorial video)

This part concludes the 1st video tutorial

+++ desktop/src/io/wasin/blockbunny/desktop/DesktopLauncher.kt
@@ -9,8 +9,8 @@ object DesktopLauncher {
     fun main(arg: Array<String>) {
         val config = LwjglApplicationConfiguration()
         config.title = Game.TITLE
-        config.width = Game.V_WIDTH * Game.SCALE
-        config.height = Game.V_HEIGHT * Game.SCALE
+        config.width = Game.V_WIDTH.toInt() * Game.SCALE
+        config.height = Game.V_HEIGHT.toInt() * Game.SCALE
         LwjglApplication(Game(), config)
     }
 }
